# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ccShield backend",
    "description": "Bilibili live-room moderator tool — backend API.",
    "version": "2.0.4"
  },
  "paths": {
...
    "/api/ban": {
      "post": {
        "tags": [
          "ban"
        ],
        "summary": "Ban a user in the given room",
        "description": "Ban ``body.uid`` in ``body.room_id`` for ``body.hour`` hours.\n\nRequires an authenticated session. On success, updates the live\n:class:`BanListManager` if it is currently running for the same\nroom so every connected WS client receives a ``ban_added`` delta.",
        "operationId": "post_ban_route_api_ban_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ban"
        ],
        "summary": "Unban a user by block id",
        "description": "Unban ``body.uid`` in ``body.room_id`` using ``body.block_id``.\n\nRequires an authenticated session. On success, broadcasts a\n``ban_removed`` delta through the running :class:`BanListManager`.",
        "operationId": "delete_ban_route_api_ban_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ban-list/{room_id}": {
      "get": {
        "tags": [
          "ban"
        ],
        "summary": "Read the current ban list for a room",
        "description": "Return the ban list for ``room_id``.\n\nReads from the running :class:`BanListManager`'s local state when\nit is currently started for ``room_id`` (the cheap path — no\nnetwork round-trip). Falls back to\n:meth:`BilibiliClient.get_ban_list` otherwise (a fresh fetch\npaginates up to ``_BAN_LIST_MAX_PAGES``).",
        "operationId": "get_ban_list_route_api_ban_list__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BanListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BanEntryResponse": {
        "properties": {
          "block_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "uid": {
            "type": "integer",
            "title": "Uid"
          },
          "uname": {
            "type": "string",
            "title": "Uname"
          },
          "operator_uid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Uid"
          },
          "operator_name": {
            "type": "string",
            "title": "Operator Name"
          },
          "hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hour"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "pending": {
            "type": "boolean",
            "title": "Pending"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "block_id",
          "uid",
          "uname",
          "operator_uid",
          "operator_name",
          "hour",
          "reason",
          "created_at",
          "expires_at",
          "pending"
        ],
        "title": "BanEntryResponse",
        "description": "Stable public shape for one B站 silent-user record."
      },
      "BanListResponse": {
        "properties": {
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "bans": {
            "items": {
              "$ref": "#/components/schemas/BanEntryResponse"
            },
            "type": "array",
            "title": "Bans"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "bans"
        ],
        "title": "BanListResponse",
        "description": "Response body for ``GET /api/ban-list/{room_id}``.\n\nRaw B站 aliases are normalized by :mod:`app.room.banlist` before\nthey cross this boundary, so every browser receives the same typed\nfields and can safely gate 解禁 on ``block_id`` / ``pending``."
      },
      "BanRequest": {
        "properties": {
          "room_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Room Id"
          },
          "uid": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Uid"
          },
          "hour": {
            "type": "integer",
            "enum": [
              -1,
              0,
              2,
              4,
              24,
              168
            ],
            "title": "Hour"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "title": "Reason",
            "default": ""
          },
          "uname": {
            "type": "string",
            "maxLength": 100,
            "title": "Uname",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "room_id",
          "uid",
          "hour"
        ],
        "title": "BanRequest",
        "description": "Request body for ``POST /api/ban``.\n\n``extra=\"forbid\"`` so a typo'd field (e.g. ``hours``) is rejected\nwith 422 instead of silently ignored — moderator actions must not\nsilently drop a parameter the operator thinks they sent."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OkResponse",
        "description": "Response body for ``POST`` / ``DELETE /api/ban`` on success."
      },
...
      "UnbanRequest": {
        "properties": {
          "room_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Room Id"
          },
          "block_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Block Id"
          },
          "uid": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Uid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "room_id",
          "block_id",
          "uid"
        ],
        "title": "UnbanRequest",
        "description": "Request body for ``DELETE /api/ban``."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}